3.1.4

+++ Ovning3/Person.cs
@@ -12,11 +12,12 @@ namespace Ovning3
         private int age;
         private string fName;
         private string lName;
-        private int height;
-        private int weight;
+        private double height;
+        private double weight;
 
-        public Person(int age, string fName, string lName, int height, int weight)
+        public Person(int age, string fName, string lName, double height, double weight)
         {
+            //Age = age;
             Age = age;
             FName = fName;
             LName = lName;
@@ -80,8 +81,8 @@ namespace Ovning3
                     lName = value;
             } 
         }
-        public int Height { get { return height; } set { height = value; } }
-        public int Weight { get { return weight; } set { weight = value; } }
+        public double Height { get { return height; } set { height = value; } }
+        public double Weight { get { return weight; } set { weight = value; } }
 
     }
 }
